Fix error

+++ cogs/moderation.py
@@ -19,7 +19,7 @@ class Moderation(commands.Cog):
 
     @commands.Cog.listener()
     async def on_message(self, message):
-        if message.author.id in UWU_KIDS:
+        if message.author.id in UWU_KIDS and len(message.attachments) != 0:
             await message.delete()
